Layout_TestCase/http/tests/xmlhttprequest/access-control-basic-denied-preflight-cache-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/access-control-basic-denied.cgi. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
Test async xhr preflight cache denial. If this test passes, there should be a single PASS below.

PASS
